backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
Pillow==9.5.0
piexif==1.1.3
mysql-connector-python==8.0.33
requests==2.31.0